separate out vlm run specs
