+++ src/helm/benchmark/presentation/test_run_entry.py
@@ -1,7 +1,9 @@
 import os
+
 from helm.common.object_spec import parse_object_spec
 from helm.benchmark.presentation.run_entry import read_run_entries
 from helm.benchmark.run_specs import construct_run_specs
+from helm.benchmark.vlm_run_specs import *  # noqa
 
 
 def test_read_all_specs():
